reorganize shared C utils
