+++ fract4d/c/model/image.h
@@ -159,9 +159,4 @@ public:
     void clear();
 };
 
-extern "C"
-{
-    void image_lookup(void *im, double x, double y, double *pr, double *pg, double *pb);
-}
-
 #endif
